feat: migrate GQL from apollo to hono
