+++ .zed/settings.json
@@ -51,5 +51,5 @@
 		"oxfmt",
 		"vtsls",
 		"json-language-server"
-	]
+	],
 }
